requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
plotly==5.18.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
PyYAML==6.1.0